refactor: extract comfy-cli human-output parsing into clitext.py
Moves five clusters from server.py into a new leaf module,
src/comfy_mcp/clitext.py, byte-for-byte: the Saved:-block / install-
failure text extractors and plain_ok synthesis (_extract_saved_paths,
_extract_install_failures, _synthesize_plain_result, _plain_message
and their supporting constants/helpers), the missing-verb/-option
capability probes (_is_missing_verb_error, _is_missing_option_error,
_normalize_cli_text), install_node's per-pack verdict
(_classify_install_result), and the echoed-argv forgery guards
(_phrase_is_only_the_caller_s, _is_manager_missing_error).

_extract_install_failures and _extract_saved_paths are the AGENTS.md-
documented cm-cli contract; verified byte-for-byte identical to the
pre-move source via ast.get_source_segment comparison. server.py
reaches every moved name module-qualified (clitext._x) at every call
site and prose cross-reference — 46 sites across ten tool functions
and their docstrings — with zero behavior re-exports.

unicodedata and Iterable dropped from server.py's imports once ruff
flagged them unused post-move (both were needed only by the moved
code, now imported by clitext.py instead). Four test files rename 21
read-only references to the clitext module; zero monkeypatches.
AGENTS.md's module table gains a clitext.py row ("Five leaf modules"
-> "Six").
