+++ tests/conftest.py
@@ -122,7 +122,7 @@ def _clear_comfyui_target_env(monkeypatch):
 def _clear_columns_env(monkeypatch):
     """Default every test to rich's own off-a-TTY console width (80 columns).
 
-    ``server._child_console_width`` reads ``COLUMNS`` because ``_comfy_env``
+    ``clitext._child_console_width`` reads ``COLUMNS`` because ``_comfy_env``
     forwards it to the child, so a developer running under a terminal that
     exports it would shift where ``_extract_saved_paths`` believes comfy-cli
     folded its output — the same class of ambient-environment perturbation the
