+++ tests/test_install_node.py
@@ -57,7 +57,7 @@ from mcp.server.elicitation import (
     DeclinedElicitation,
 )
 
-from comfy_mcp import instructions, server
+from comfy_mcp import clitext, instructions, server
 
 # Captured before any test patches it, so the one test that wants the REAL
 # `comfy env` pre-flight can put it back and drive the probe end to end through
@@ -587,7 +587,7 @@ def test_the_sentence_still_has_to_name_something():
     pip log prints about "installing" a wheel flip an install to `ok: False`.
     """
     assert (
-        server._extract_install_failures(
+        clitext._extract_install_failures(
             "ERROR: An error occurred while installing the wheel for torch\n"
         )
         == []
@@ -615,13 +615,13 @@ def test_a_flood_on_one_stream_cannot_evict_the_real_failure(patched_plain_run):
     flood = "".join(
         f"ERROR: An error occurred while installing 'noise-{index}'.\n"
         "Failed to clone repository\n"
-        for index in range(server._MAX_INSTALL_FAILURES * 2)
+        for index in range(clitext._MAX_INSTALL_FAILURES * 2)
     )
     patched_plain_run(0, stdout=_NOT_FOUND_OUTPUT, stderr=flood)
 
     result = _install(["no-such-pack-xyz"], ctx=_FakeCtx())
 
-    assert len(result["result"]["failures"]) == server._MAX_INSTALL_FAILURES
+    assert len(result["result"]["failures"]) == clitext._MAX_INSTALL_FAILURES
     assert result["installed"] == []
     assert result["failed"][0]["name"] == "no-such-pack-xyz"
 
@@ -726,7 +726,7 @@ def test_an_unreadable_failures_field_degrades_to_the_success_payload(failures):
     pack(s) failed" alongside an empty list would be a worse lie than the one this
     replaced.
     """
-    result = server._classify_install_result(
+    result = clitext._classify_install_result(
         ["comfyui-impact-pack"], {"ok": True, "failures": failures}
     )
 
@@ -742,7 +742,7 @@ def test_a_record_with_unusable_fields_still_counts_as_a_failure():
     reported (with an empty name) rather than dropped — the same one-directional
     choice the extractor makes when cm-cli's sentence carried no usable name.
     """
-    result = server._classify_install_result(
+    result = clitext._classify_install_result(
         ["comfyui-impact-pack"],
         {"ok": False, "failures": [{"pack": 7, "reason": None}]},
     )
